Moved clap to main and fixed some annotation
